Fix end tracking for implicit accumulation variables.
If the last link is unknown, also check whether the accumulation variable has
been modified.  This can happen when some commands modify the beginning of the
list while others modify then end.  If modifying the beginning while the
variable is nil, then the command sets the variable without updating the
location of the last link.