Fix using `setf` on `plist-get` on Emacs versions less than 28.
This feature was added to Emacs 28 in September 2020, and we use it for the
`&key` variables in the new destructuring system.

The definition of that feature has been copied into `loopy-misc.el`, and will be
used on Emacs versions less than 28 if a setter function for `plist-get` is not
found.

For clarity, we `require` `gv.el` in a few files, but that probably isn't
needed.