Remove warning about destructuring from `map`.
Previously, this command used proper lists for destructuring, but now uses
dotted lists.  We added a warning for those still using proper lists,
but now, with things like `(key &rest val)`, it is possible to correctly
destructure using proper lists.

Instead of trying to confirm the correctness of the argument, it is better to
just remove the warning.