Add a `:test` to numbers.  Deprecate the non-keyword arguments.
- Modify `numbers`, `numbers-down`, and `numbers-up`.
- Fix signalling an error when arguments which give directions conflict.
- Add more tests.
- Update docs, changelog, and readme.

These arguments were meant to be similar to the arguments of Pythons `range`
feature, but, depending on prior knowledge, would generally produce worse code.
Cases in which the direction of the iteration (up or down) is unknown can now be
handled by the new `:test` argument, which is more flexible than the non-keyword
arguments anyway.