Bump to version 0.12.2 to correct the `.elpaignore` file. (#192)
Remove `loopy-pkg.el` from the file `.elpaignore`.  The file `.elpaignore` file
determines what goes into the tarball, not what is downloaded.  The repo
contains a file `loopy-pkg.el`, but the ELPA process also makes its own, which
should be included in the generated tarball so that the tarball can be installed
properly.