Test installing Info docs and add more Makefile targets.
Now test installing the Info documentation using the programs `makeinfo` and
`install-info`.

For some reason, we need to copy the files manually into the installation
directly.  `pacakge-install-buffer` will not do it, and the new Makefile target
for tar files does not currently work.

Instead of manually adding files to the load path, we activate the packages.