Generate `loopy-pkg.el` from the main file. (#225)
In PR #224, it was reported that `loopy-pkg.el` was out of date.

Instead of keeping it in the repo, create it for testing purposes in the Makefile. The file is created using the built-in library `lisp-mnt`, as done in `elpa-admin.el` in the ELPA repo. The generated `loopy-pkg.el` is a simpler version that is sufficient for testing installation but does not contain all of the metadata captured by what ELPA and MELPA do.