Allow `finally-do` to modify an implied `loopy-result`. (#244)
Use of the `prog1` was not allowing `finally-do` to modify the implied
return value of a loop stored in `loopy-result`.  Conceptually, an implied
`loopy-result` should not be different from an explicit `loopy-result` in this
context, outside of the variable value needing to be finalized.

The new expansion of the macro is still consistent with the example
expansion given in `loopy--expand-to-loop`.

Previously, one would've needed to use `finally-return` after the
`finally-do`:

```emacs-lisp
;; => (0 1 2 3)
(loopy (list i '(1 2 3))
       (collect loopy-result i)
       (finally-do (push 0 loopy-result))
       (finally-return loopy-result))
```

Now it works more like expected:

```emacs-lisp
;; => (0 1 2 3)
(loopy (list i '(1 2 3))
       (collect i)
       (finally-do (push 0 loopy-result)))
```

- Add tests:
  - `finally-do-modifies-implied-result-var`
  - `finally-do-modifies-implied-result-var-early-return`

- Update Org doc:
  - on how `finally-do` can interact with variables.
  - Clarify that `finally-do` is run before `finally-return`.