Don't use a completion check when `finally-return` is used. (#249)
In PR #244, we added a completion test when using `finally-do` in case
the loop would have used `loopy-result` and maybe modified it in `finally-do`,
but the loop was apparently terminated early, in which case we
should have used the value captured by the `cl-block`.

When we added that, we incorrectly set the variable
when using `finally-do` even when we were using `finally-return`,
in which case the completion test in unneeded.