For destructuring accumulation, maintain order of expressions used by `pcase`. (#251)
When destructuring for accumulation commands, don't assume that `pcase` binds
the final variable to a value already stored in a temporary variable.  More
specifically, in `loopy--pcase-parse-for-destructuring-accumulation-command`
keep the same order returned by `pcase` for the binding of those final
variables.

This is not known to be an error, but could possibly produce one for some
unknown `pcase` pattern.